Conflicts:
	MedicalCare/src/main/java/ui/dailysheetfilling/listeners/ListenerLotNumber.java

+++ MedicalCare/src/main/java/tests/DailyTest.java
@@ -278,7 +278,7 @@ public class DailyTest extends Analysis {
          * @param P
          * @param p
 	 */
-	public void addBlood(float Hb, float GR, float GB, float hemato, float P, Patient p) throws Exception {
+	public void addBlood(int Hb, int GR, int GB, int hemato, int P, Patient p) throws Exception {
 		this.bloodTest = new BloodTest(Hb, GR, GB, hemato, P, p);
 	}
